Sourcing env.sh will not any more preselect a docker recipe. The latter is not done with the command drecipe.

+++ bin/build.sh
@@ -5,6 +5,13 @@
 #flavor=${1}
 #shift
 
+if [ ! -r /tmp/dev-scripts-recipe-dir-$PPID ]
+then
+  echo unknown recipe
+  exit
+fi
+export DEV_SCRIPTS_DOCKER_DIR=`cat /tmp/dev-scripts-recipe-dir-$PPID`
+
 cd ${DEV_SCRIPTS_DOCKER_DIR}
 
 rm -rf mydevtools
